requirements.txt
# ============================================================================
# Trading Platform API - Python Dependencies
# ============================================================================

# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# API Clients & HTTP
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# JWT & Authentication
PyJWT==2.8.1
cryptography==41.0.7

# OpenAI
openai==1.6.1

# GitHub
PyGithub==2.1.1

# Pinecone
pinecone-client==3.0.1

# Binance
python-binance==1.0.17

# Coinbase
# Note: Using requests for API calls; no official Python SDK for Advanced Trade API

# Trading & Finance
yfinance==0.2.32
pandas==2.1.4
numpy==1.26.3
TA-Lib==0.4.28

# Discord & Slack
discord.py==2.3.2
slack-sdk==3.26.1

# Async & Concurrency
asyncio-contextmanager==1.0.0
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2

# Web Framework (Optional - for FastAPI examples)
fastapi==0.109.0
uvicorn==0.27.0
starlette==0.35.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Development
ipython==8.19.0
jupyter==1.0.0
notebook==7.0.6
